Resolve comments in package runtime split
