+++ internal/controller/pkg/manager/reconciler_test.go
@@ -1051,264 +1051,3 @@ func TestReconcile(t *testing.T) {
 		})
 	}
 }
-
-func TestPackageHealth(t *testing.T) {
-	type args struct {
-		pr v1.PackageRevision
-	}
-	type want struct {
-		condition commonv1.Condition
-	}
-
-	cases := map[string]struct {
-		reason string
-		args   args
-		want   want
-	}{
-		"HealthyRevisionAndRuntimeWithRuntime": {
-			reason: "Should return healthy condition when both revision and runtime are healthy for package with runtime",
-			args: args{
-				pr: &v1.ProviderRevision{
-					Status: v1.PackageRevisionStatus{
-						ConditionedStatus: commonv1.ConditionedStatus{
-							Conditions: []commonv1.Condition{
-								{
-									Type:    v1.TypeRevisionHealthy,
-									Status:  corev1.ConditionTrue,
-									Reason:  v1.ReasonHealthy,
-									Message: "Package revision is healthy",
-								},
-								{
-									Type:    v1.TypeRuntimeHealthy,
-									Status:  corev1.ConditionTrue,
-									Reason:  v1.ReasonHealthy,
-									Message: "Package runtime is healthy",
-								},
-							},
-						},
-					},
-				},
-			},
-			want: want{
-				condition: v1.Healthy(),
-			},
-		},
-		"HealthyRevisionWithoutRuntime": {
-			reason: "Should return healthy condition when revision is healthy for package without runtime",
-			args: args{
-				pr: &v1.ConfigurationRevision{
-					Status: v1.PackageRevisionStatus{
-						ConditionedStatus: commonv1.ConditionedStatus{
-							Conditions: []commonv1.Condition{
-								{
-									Type:    v1.TypeRevisionHealthy,
-									Status:  corev1.ConditionTrue,
-									Reason:  v1.ReasonHealthy,
-									Message: "Package revision is healthy",
-								},
-							},
-						},
-					},
-				},
-			},
-			want: want{
-				condition: v1.Healthy(),
-			},
-		},
-		"UnhealthyRevision": {
-			reason: "Should return unhealthy condition when revision is unhealthy",
-			args: args{
-				pr: &v1.ProviderRevision{
-					Status: v1.PackageRevisionStatus{
-						ConditionedStatus: commonv1.ConditionedStatus{
-							Conditions: []commonv1.Condition{
-								{
-									Type:    v1.TypeRevisionHealthy,
-									Status:  corev1.ConditionFalse,
-									Reason:  v1.ReasonUnhealthy,
-									Message: "Package revision is not ready",
-								},
-								{
-									Type:    v1.TypeRuntimeHealthy,
-									Status:  corev1.ConditionTrue,
-									Reason:  v1.ReasonHealthy,
-									Message: "Package runtime is healthy",
-								},
-							},
-						},
-					},
-				},
-			},
-			want: want{
-				condition: v1.Unhealthy().WithMessage("Package revision health is \"False\" with message: Package revision is not ready"),
-			},
-		},
-		"UnhealthyRuntime": {
-			reason: "Should return unhealthy condition when runtime is unhealthy",
-			args: args{
-				pr: &v1.ProviderRevision{
-					Status: v1.PackageRevisionStatus{
-						ConditionedStatus: commonv1.ConditionedStatus{
-							Conditions: []commonv1.Condition{
-								{
-									Type:    v1.TypeRevisionHealthy,
-									Status:  corev1.ConditionTrue,
-									Reason:  v1.ReasonHealthy,
-									Message: "Package revision is healthy",
-								},
-								{
-									Type:    v1.TypeRuntimeHealthy,
-									Status:  corev1.ConditionFalse,
-									Reason:  v1.ReasonUnhealthy,
-									Message: "Runtime deployment is not ready",
-								},
-							},
-						},
-					},
-				},
-			},
-			want: want{
-				condition: v1.Unhealthy().WithMessage("Package runtime health is \"False\" with message: Runtime deployment is not ready"),
-			},
-		},
-		"BothUnhealthy": {
-			reason: "Should return unhealthy condition with revision message when both are unhealthy (revision checked first)",
-			args: args{
-				pr: &v1.FunctionRevision{
-					Status: v1.FunctionRevisionStatus{
-						PackageRevisionStatus: v1.PackageRevisionStatus{
-							ConditionedStatus: commonv1.ConditionedStatus{
-								Conditions: []commonv1.Condition{
-									{
-										Type:    v1.TypeRevisionHealthy,
-										Status:  corev1.ConditionFalse,
-										Reason:  v1.ReasonUnhealthy,
-										Message: "Package revision is not ready",
-									},
-									{
-										Type:    v1.TypeRuntimeHealthy,
-										Status:  corev1.ConditionFalse,
-										Reason:  v1.ReasonUnhealthy,
-										Message: "Runtime deployment is not ready",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			want: want{
-				condition: v1.Unhealthy().WithMessage("Package revision health is \"False\" with message: Package revision is not ready"),
-			},
-		},
-		"UnknownRevisionHealth": {
-			reason: "Should return unhealthy condition when revision health is unknown",
-			args: args{
-				pr: &v1.ProviderRevision{
-					Status: v1.PackageRevisionStatus{
-						ConditionedStatus: commonv1.ConditionedStatus{
-							Conditions: []commonv1.Condition{
-								{
-									Type:   v1.TypeRevisionHealthy,
-									Status: corev1.ConditionUnknown,
-									Reason: v1.ReasonUnknownHealth,
-								},
-								{
-									Type:    v1.TypeRuntimeHealthy,
-									Status:  corev1.ConditionTrue,
-									Reason:  v1.ReasonHealthy,
-									Message: "Package runtime is healthy",
-								},
-							},
-						},
-					},
-				},
-			},
-			want: want{
-				condition: v1.Unhealthy().WithMessage("Package revision health is \"Unknown\""),
-			},
-		},
-		"UnknownRuntimeHealth": {
-			reason: "Should return unhealthy condition when runtime health is unknown for package with runtime",
-			args: args{
-				pr: &v1.ProviderRevision{
-					Status: v1.PackageRevisionStatus{
-						ConditionedStatus: commonv1.ConditionedStatus{
-							Conditions: []commonv1.Condition{
-								{
-									Type:    v1.TypeRevisionHealthy,
-									Status:  corev1.ConditionTrue,
-									Reason:  v1.ReasonHealthy,
-									Message: "Package revision is healthy",
-								},
-								{
-									Type:   v1.TypeRuntimeHealthy,
-									Status: corev1.ConditionUnknown,
-									Reason: v1.ReasonUnknownHealth,
-								},
-							},
-						},
-					},
-				},
-			},
-			want: want{
-				condition: v1.Unhealthy().WithMessage("Package runtime health is \"Unknown\""),
-			},
-		},
-		"MissingConditions": {
-			reason: "Should return unhealthy condition when conditions are missing",
-			args: args{
-				pr: &v1.ConfigurationRevision{
-					Status: v1.PackageRevisionStatus{
-						ConditionedStatus: commonv1.ConditionedStatus{
-							Conditions: []commonv1.Condition{},
-						},
-					},
-				},
-			},
-			want: want{
-				condition: v1.Unhealthy().WithMessage("Package revision health is \"Unknown\""),
-			},
-		},
-		"FunctionRevisionHealthy": {
-			reason: "Should return healthy condition for function revision with both conditions healthy",
-			args: args{
-				pr: &v1.FunctionRevision{
-					Status: v1.FunctionRevisionStatus{
-						PackageRevisionStatus: v1.PackageRevisionStatus{
-							ConditionedStatus: commonv1.ConditionedStatus{
-								Conditions: []commonv1.Condition{
-									{
-										Type:    v1.TypeRevisionHealthy,
-										Status:  corev1.ConditionTrue,
-										Reason:  v1.ReasonHealthy,
-										Message: "Function revision is healthy",
-									},
-									{
-										Type:    v1.TypeRuntimeHealthy,
-										Status:  corev1.ConditionTrue,
-										Reason:  v1.ReasonHealthy,
-										Message: "Function runtime is healthy",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			want: want{
-				condition: v1.Healthy(),
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			got := packageHealth(tc.args.pr)
-
-			if diff := cmp.Diff(tc.want.condition, got); diff != "" {
-				t.Errorf("\n%s\npackageHealth(...): -want, +got:\n%s", tc.reason, diff)
-			}
-		})
-	}
-}
